refactor(preset-mini): expose rules in larger block
