+++ packages/preset-mini/src/rules/border.ts
@@ -1,34 +1,30 @@
-import { CSSEntries, Rule, RuleContext } from '@unocss/core'
+import { CSSEntries, RuleContext, Rule } from '@unocss/core'
 import { Theme } from '../theme'
 import { cornerMap, directionMap, handler as h } from '../utils'
-import { borderColors } from './color'
+import { colorResolver } from './color'
 
-export const borderSizes: Rule[] = [
+export const borders: Rule[] = [
+  // size
   [/^border$/, handlerBorder],
   [/^(?:border|b)(?:-([^-]+))?$/, handlerBorder],
   [/^(?:border|b)(?:-([^-]+))?(?:-([^-]+))?$/, handlerBorder],
-]
 
-export const borderRadius: Rule[] = [
-  [/^(?:border-)?(?:rounded|rd)$/, handlerRounded],
-  [/^(?:border-)?(?:rounded|rd)(?:-([^-]+))?$/, handlerRounded],
-  [/^(?:border-)?(?:rounded|rd)(?:-([^-]+))?(?:-([^-]+))?$/, handlerRounded],
-]
-
-export const borderStyles: Rule[] = [
+  // style
   ['border-solid', { 'border-style': 'solid' }],
   ['border-dashed', { 'border-style': 'dashed' }],
   ['border-dotted', { 'border-style': 'dotted' }],
   ['border-double', { 'border-style': 'double' }],
   ['border-none', { 'border-style': 'none' }],
-]
 
-export const borders = [
-  borderSizes,
-  borderColors,
-  borderStyles,
-  borderRadius,
-].flat(1)
+  // radius
+  [/^(?:border-)?(?:rounded|rd)$/, handlerRounded],
+  [/^(?:border-)?(?:rounded|rd)(?:-([^-]+))?$/, handlerRounded],
+  [/^(?:border-)?(?:rounded|rd)(?:-([^-]+))?(?:-([^-]+))?$/, handlerRounded],
+
+  // colors
+  [/^(?:border|b)-(.+)$/, colorResolver('border-color', 'border')],
+  [/^(?:border|b)-op(?:acity)?-?(.+)$/, ([, opacity]) => ({ '--un-border-opacity': h.bracket.percent(opacity) })],
+]
 
 function handlerBorder([, a, b]: string[]): CSSEntries | undefined {
   const [d, s = '1'] = directionMap[a] ? [a, b] : ['', a]
