+++ packages/preset-wind/src/rules/default.ts
@@ -1,8 +1,6 @@
 import { Rule } from '@unocss/core'
 import {
-  alignContents,
-  alignItems,
-  alignSelfs,
+  alignments,
   appearance,
   appearances,
   aspectRatio,
@@ -17,27 +15,20 @@ import {
   fillColors,
   flex,
   floats,
-  fontSizes,
   fontSmoothings,
   fontStyles,
-  fontWeights,
-  fontsFamilies,
+  fonts,
   gaps,
   grids,
   insets,
   justifies,
-  justifyItems,
-  justifySelfs,
-  leadings,
   margins,
   opacity,
   orders,
   outline,
   overflows,
   paddings,
-  placeContents,
-  placeItems,
-  placeSelfs,
+  placements,
   placeholder,
   pointerEvents,
   positions,
@@ -49,19 +40,15 @@ import {
   textAligns,
   textColors,
   textDecorations,
-  textDecorationColors,
   textIndents,
   textOverflows,
   textShadows,
-  textStrokeColors,
-  textStrokeWidths,
-  trackings,
+  textStrokes,
   transforms,
   transitions,
   userSelects,
   verticalAligns,
   whitespaces,
-  wordSpacings,
   zIndexes,
 } from '@unocss/preset-mini/rules'
 import { container } from './container'
@@ -103,16 +90,12 @@ export const rules: Rule[] = [
   borders,
   divides,
   contents,
-  fontsFamilies,
-  fontSizes,
-  fontWeights,
+  fonts,
   tabSizes,
   textIndents,
   textOverflows,
   textDecorations,
-  textDecorationColors,
-  textStrokeWidths,
-  textStrokeColors,
+  textStrokes,
   textShadows,
   textTransforms,
   textAligns,
@@ -144,9 +127,6 @@ export const rules: Rule[] = [
   caretOpacity,
   imageRenderings,
   breaks,
-  trackings,
-  wordSpacings,
-  leadings,
   overflows,
   outline,
   appearance,
@@ -155,14 +135,8 @@ export const rules: Rule[] = [
   positions,
   orders,
   justifies,
-  justifyItems,
-  justifySelfs,
-  alignContents,
-  alignItems,
-  alignSelfs,
-  placeContents,
-  placeItems,
-  placeSelfs,
+  placements,
+  alignments,
   animations,
   insets,
   floats,
